add PG books rsync-and-store-in-DB process via a Python script
