+++ server/.docker/python/Dockerfile
@@ -3,3 +3,5 @@ FROM python:3.6-alpine
 RUN apk update \
   && apk add gcc python3-dev musl-dev \
   && apk add postgresql-dev
+
+RUN apk add rsync
