+++ server/docker-compose.yml
@@ -37,7 +37,7 @@ services:
   nginx:
     image: nginx:1-alpine
     volumes:
-      - ./docker/nginx/nginx.conf:/etc/nginx/conf.d/default.conf:ro
+      - ./.docker/nginx/nginx.conf:/etc/nginx/conf.d/default.conf:ro
       - ~/gutenberg-mirror/generated-collection:/www/gutenberg-mirror/generated-collection:ro
     depends_on:
       - postgrest
@@ -55,7 +55,7 @@ services:
       - 8084:80
 
   postgrest:
-    build: docker/postgrest
+    build: .docker/postgrest
     volumes:
       - ./postgrest-api-public.conf:/etc/postgrest.conf
     ports:
